fix: send selected columns to analytics
