+++ internal/lookoutui/src/analytics/index.ts
@@ -1,4 +1,5 @@
 export * from "./Analytics"
 export * from "./AnalyticsScript"
 export * from "./AnalyticsUserIdentifier"
+export * from "./trackAnalyticsEvent"
 export * from "./types"
